if author doesnt match user redirect + flash error

+++ YelpCamp/routes/campgrounds.js
@@ -51,15 +51,29 @@ router.get('/:id', wrapAsync(async (req, res) => {
 
 // Edit path/route form
 router.get('/:id/edit', isLoggedIn, wrapAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id)
+    const {id} = req.params
+    const campground = await Campground.findById(id)
+    if(!campground){
+        req.flash('error','Cannot find that campground!')
+        return res.redirect('/campgrounds')
+    }
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error','You are not authorized to edit someone elses post!')
+        return res.redirect(`/campgrounds/${id}`)
+    }
     res.render(`campgrounds/edit`, { campground })
 }))
 
 // Path to update edit information on campground
-router.put('/:id', validateCampground, wrapAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, validateCampground, wrapAsync(async (req, res) => {
     const { id } = req.params
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error','Unauthorized action, unable to edit someone elses post')
+        return res.redirect(`/campgrounds/${id}`)
+    }
     // take spread of object from req.body.campground in edit.ejs form
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground })
+    const campEdit = await Campground.findByIdAndUpdate(id, { ...req.body.campground })
     req.flash('success', 'Succesfully updated campground')
     res.redirect(`/campgrounds/${campground._id}`)
 }))
@@ -67,7 +81,13 @@ router.put('/:id', validateCampground, wrapAsync(async (req, res) => {
 // Delete path/route
 router.delete('/:id',isLoggedIn, wrapAsync(async (req, res) => {
     const { id } = req.params
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)){
+        req.flash('error','You do not have permission to delete someone elses post!')
+        return res.redirect(`/campgrounds/${id}`)
+    }
     await Campground.findByIdAndDelete(id)
+    req.flash('success','Succesfully deleted campground!')
     res.redirect('/campgrounds')
 }))
